Rename / Reorder bytecode tests
