+++ Compiler/test2/cc_bytecode_test_0.cpp
@@ -251,167 +251,7 @@ TEST_F(Bytecode0, NotNot) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, SimpleIntFunction) { 
-
-    char *inpl = "\
-        int Foo()      \n\
-    {                  \n\
-        return 15;     \n\
-    }";
-   
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("SimpleIntFunction", scrip);
-    const size_t codesize = 11;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,           15,   31,    3,    6,    // 7
-       3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-
-}
-
-TEST_F(Bytecode0, IntFunctionLocalV) {
-
-    char *inpl = "\
-        int Foo()       \n\
-        {               \n\
-            int a = 15; \n\
-            return a;   \n\
-        }";
-
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("IntFunctionLocalV", scrip);
-    const size_t codesize = 23;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,           15,   29,    3,   51,    // 7
-       4,    7,    3,    2,            1,    4,   31,    6,    // 15
-       2,    1,    4,    6,            3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, IntFunctionParam) {
-
-    char *inpl = "\
-        int Foo(int a) \n\
-    {                  \n\
-        return a;      \n\
-    }";
-    
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("IntFunctionParam", scrip);
-    const size_t codesize = 12;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   51,    8,            7,    3,   31,    3,    // 7
-       6,    3,    0,    5,          -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, IntFunctionGlobalV) {    
-
-    char *inpl = "\
-        int a = 15;    \n\
-        int Foo( )     \n\
-    {                  \n\
-        return a;      \n\
-    }";
-   
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("IntFunctionGlobalV", scrip);
-    const size_t codesize = 13;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    2,            0,    7,    3,   31,    // 7
-       3,    6,    3,    0,            5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 1;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-       4,  -999
-    };
-    char fixuptypes[] = {
-      1,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Float1) {   
+TEST_F(Bytecode0, Float01) {   
 
     // Float values
 
@@ -436,7 +276,7 @@ TEST_F(Bytecode0, Float1) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("Float1", scrip);
+    // WriteOutput("Float01", scrip);
     const size_t codesize = 156;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -480,7 +320,7 @@ TEST_F(Bytecode0, Float1) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, Float2) {   
+TEST_F(Bytecode0, Float02) {   
 
     // Positive and negative float parameter defaults
 
@@ -499,7 +339,7 @@ TEST_F(Bytecode0, Float2) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("Float2", scrip);
+    // WriteOutput("Float02", scrip);
     const size_t codesize = 65;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -540,7 +380,7 @@ TEST_F(Bytecode0, Float2) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, FloatExpr1) { 
+TEST_F(Bytecode0, Float03) { 
 
     char *inpl = "\
         float a = 15.0;     \n\
@@ -554,7 +394,7 @@ TEST_F(Bytecode0, FloatExpr1) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("FloatExpr1", scrip);
+    // WriteOutput("Float03", scrip);
     const size_t codesize = 38;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -591,7 +431,7 @@ TEST_F(Bytecode0, FloatExpr1) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, FloatExpr2) {    
+TEST_F(Bytecode0, Float04) {    
 
     char *inpl = "\
         float a = 15.0;                             \n\
@@ -609,7 +449,7 @@ TEST_F(Bytecode0, FloatExpr2) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("FloatExpr2", scrip);
+    // WriteOutput("Float04", scrip);
     const size_t codesize = 244;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -672,7 +512,7 @@ TEST_F(Bytecode0, FloatExpr2) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, IfThenElse1) {  
+TEST_F(Bytecode0, FlowIfThenElse1) {
 
     char *inpl = "\
     int Foo()               \n\
@@ -688,7 +528,7 @@ TEST_F(Bytecode0, IfThenElse1) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("IfThenElse1", scrip);
+    // WriteOutput("FlowIfThenElse1", scrip);
     const size_t codesize = 102;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -725,7 +565,7 @@ TEST_F(Bytecode0, IfThenElse1) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, IfThenElse2) {    
+TEST_F(Bytecode0, FlowIfThenElse2) {
 
     char *inpl = "\
     int Foo()               \n\
@@ -742,7 +582,7 @@ TEST_F(Bytecode0, IfThenElse2) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("IfThenElse2", scrip);
+    // WriteOutput("FlowIfThenElse2", scrip);
     const size_t codesize = 102;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -844,7 +684,7 @@ TEST_F(Bytecode0, While) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, DoNCall) {
+TEST_F(Bytecode0, FlowDoNCall) {
 
     char *inpl = "\
     char c = 'x';             \n\
@@ -869,7 +709,7 @@ TEST_F(Bytecode0, DoNCall) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("DoNCall", scrip);
+    // WriteOutput("FlowDoNCall", scrip);
     const size_t codesize = 120;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -917,7 +757,7 @@ TEST_F(Bytecode0, DoNCall) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, DoUnbracedIf) {    
+TEST_F(Bytecode0, FlowDoUnbracedIf) {    
 
     char *inpl = "\
     void noloopcheck main()   \n\
@@ -935,7 +775,7 @@ TEST_F(Bytecode0, DoUnbracedIf) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("DoUnbracedIf", scrip);
+    // WriteOutput("FlowDoUnbracedIf", scrip);
     const size_t codesize = 70;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -969,7 +809,7 @@ TEST_F(Bytecode0, DoUnbracedIf) {
 }
 
 
-TEST_F(Bytecode0, For1) {  
+TEST_F(Bytecode0, FlowFor1) {  
 
     char *inpl = "\
     int loop;                       \n\
@@ -987,7 +827,7 @@ TEST_F(Bytecode0, For1) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For1", scrip);
+    // WriteOutput("FlowFor1", scrip);
     const size_t codesize = 119;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1034,7 +874,7 @@ TEST_F(Bytecode0, For1) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, For2) {    
+TEST_F(Bytecode0, FlowFor2) {
 
     char *inpl = "\
     int Foo(int i, float f)         \n\
@@ -1060,7 +900,7 @@ TEST_F(Bytecode0, For2) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For2", scrip);
+    // WriteOutput("FlowFor2", scrip);
     const size_t codesize = 312;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1124,7 +964,7 @@ TEST_F(Bytecode0, For2) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, For3) {
+TEST_F(Bytecode0, FlowFor3) {
 
     char *inpl = "\
     managed struct Struct           \n\
@@ -1146,7 +986,7 @@ TEST_F(Bytecode0, For3) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For3", scrip);
+    // WriteOutput("FlowFor3", scrip);
     const size_t codesize = 57;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1186,7 +1026,7 @@ TEST_F(Bytecode0, For3) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, For4) {    
+TEST_F(Bytecode0, FlowFor4) {
 
     char *inpl = "\
     void main()                     \n\
@@ -1200,7 +1040,7 @@ TEST_F(Bytecode0, For4) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For4", scrip);
+    // WriteOutput("FlowFor4", scrip);
     const size_t codesize = 71;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1233,7 +1073,7 @@ TEST_F(Bytecode0, For4) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, For5) {    
+TEST_F(Bytecode0, FlowFor5) {
 
     char *inpl = "\
         int Start()                     \n\
@@ -1261,7 +1101,7 @@ TEST_F(Bytecode0, For5) {
 
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For5", scrip);
+    // WriteOutput("FlowFor5", scrip);
     const size_t codesize = 140;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1311,7 +1151,7 @@ TEST_F(Bytecode0, For5) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, For6) {  
+TEST_F(Bytecode0, FlowFor6) {
 
     char *inpl = "\
         void main()                     \n\
@@ -1340,7 +1180,7 @@ TEST_F(Bytecode0, For6) {
 
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For6", scrip);
+    // WriteOutput("FlowFor6", scrip);
     const size_t codesize = 140;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1390,7 +1230,7 @@ TEST_F(Bytecode0, For6) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, For7) {
+TEST_F(Bytecode0, FlowFor7) {
 
     // Initializer and iterator of a for() need not be assignments,
     // they can be func calls.
@@ -1422,7 +1262,7 @@ TEST_F(Bytecode0, For7) {
 
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("For7", scrip);
+    // WriteOutput("FlowFor7", scrip);
     const size_t codesize = 123;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1472,7 +1312,7 @@ TEST_F(Bytecode0, For7) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, Continue1) {   
+TEST_F(Bytecode0, FlowContinue1) {   
 
     // Locals only become invalid at the end of their nesting; below a "continue", they
     // remain valid so the offset to start of the local block
@@ -1498,7 +1338,7 @@ TEST_F(Bytecode0, Continue1) {
 
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("Continue1", scrip);
+    // WriteOutput("FlowContinue1", scrip);
     const size_t codesize = 114;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1537,7 +1377,7 @@ TEST_F(Bytecode0, Continue1) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, IfDoWhile) {   
+TEST_F(Bytecode0, FlowIfDoWhile) {
 
     char *inpl = "\
     int Foo(int i, float f)                      \n\
@@ -1557,7 +1397,7 @@ TEST_F(Bytecode0, IfDoWhile) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("IfDoWhile", scrip);
+    // WriteOutput("FlowIfDoWhile", scrip);
     const size_t codesize = 179;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1604,7 +1444,7 @@ TEST_F(Bytecode0, IfDoWhile) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
-TEST_F(Bytecode0, Switch01) {    
+TEST_F(Bytecode0, FlowSwitch01) {    
 
     char *inpl = "\
     int Foo(int i, float f)         \n\
@@ -1623,7 +1463,7 @@ TEST_F(Bytecode0, Switch01) {
     int compileResult = cc_compile(inpl, scrip);
     ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
 
-    // WriteOutput("Switch01", scrip);
+    // WriteOutput("FlowSwitch01", scrip);
     const size_t codesize = 165;
     EXPECT_EQ(codesize, scrip.codesize);
 
@@ -1649,6 +1489,69 @@ TEST_F(Bytecode0, Switch01) {
        6,    3,    5,   30,            4,   15,    3,    4,    // 151
       70,  -93,   31, -124,            6,    3,    0,   31,    // 159
        3,    6,    3,    0,            5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+
+    int idx2 = -1;
+    for (size_t idx = 0; static_cast<int>(idx) < scrip.numimports; idx++)
+    {
+        if (!strcmp(scrip.imports[idx], ""))
+            continue;
+        idx2++;
+        ASSERT_LT(idx2, numimports);
+        std::string prefix = "imports[";
+        prefix += std::to_string(idx2) + "] == ";
+        std::string is_val = prefix + scrip.imports[idx];
+        std::string test_val = prefix + imports[idx2];
+        ASSERT_EQ(is_val, test_val);
+    }
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, FlowSwitch02) {
+
+    // Last switch clause no "break"
+    char *inpl = "\
+        void main()                     \n\
+        {                               \n\
+            int i = 5;                  \n\
+            switch(i)                   \n\
+            {                           \n\
+            default: break;             \n\
+            case 5: i = 0;              \n\
+            }                           \n\
+            return;                     \n\
+        }                               \n\
+        ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("FlowSwitch02", scrip);
+    const size_t codesize = 50;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,            5,   29,    3,   51,    // 7
+       4,    7,    3,    3,            3,    4,   31,   11,    // 15
+      31,   23,    6,    3,            0,   51,    4,    8,    // 23
+       3,   31,   14,   29,            4,    6,    3,    5,    // 31
+      30,    4,   15,    3,            4,   70,  -21,   31,    // 39
+     -25,    2,    1,    4,           31,    3,    2,    1,    // 47
+       4,    5,  -999
     };
     CompareCode(&scrip, codesize, code);
 
@@ -1732,131 +1635,6 @@ TEST_F(Bytecode0, FreeLocalPtr) {
 
     const size_t stringssize = 0;
     EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, StringOldstyle01) {
-    ccSetOption(SCOPT_OLDSTRINGS, true);    
-
-    char *inpl = "\
-        int Sentinel1;              \n\
-        string GLOBAL;              \n\
-        int Sentinel2;              \n\
-                                    \n\
-        string MyFunction(int a)    \n\
-        {                           \n\
-            string x;               \n\
-            char   Sentinel3;       \n\
-            return GLOBAL;          \n\
-        }                           \n\
-        ";
-
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("StringOldstyle01", scrip);
-    const size_t codesize = 34;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-          38,    0,   51,    0,           63,  200,    1,    1,    // 7
-         200,   51,    0,   63,            1,    1,    1,    1,    // 15
-           6,    2,    4,    3,            2,    3,    2,    1,    // 23
-         201,   31,    6,    2,            1,  201,    6,    3,    // 31
-           0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 1;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      18,  -999
-    };
-    char fixuptypes[] = {
-      1,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, StringOldstyle02) {
-
-    char *inpl = "\
-        int sub(const string s) \n\
-        {                       \n\
-            return;             \n\
-        }                       \n\
-                                \n\
-        int main()              \n\
-        {                       \n\
-            sub(\"Foo\");       \n\
-        }                       \n\
-        ";
-
-    ccSetOption(SCOPT_OLDSTRINGS, true);
-
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("StringOldstyle02", scrip);
-    const size_t codesize = 30;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,            0,   31,    3,    6,    // 7
-       3,    0,    5,   38,           11,    6,    3,    0,    // 15
-      29,    3,    6,    3,            0,   23,    3,    2,    // 23
-       1,    4,    6,    3,            0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      15,   20,  -999
-    };
-    char fixuptypes[] = {
-      3,   2,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 4;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-
-    char strings[] = {
-    'F',  'o',  'o',    0,          '\0'
-    };
-
-    for (size_t idx = 0; idx < stringssize; idx++)
-    {
-        if (static_cast<int>(idx) >= scrip.stringssize) break;
-        std::string prefix = "strings[";
-        prefix += std::to_string(idx) + "] == ";
-        std::string is_val = prefix + std::to_string(strings[idx]);
-        std::string test_val = prefix + std::to_string(scrip.strings[idx]);
-        ASSERT_EQ(is_val, test_val);
-    }
 }
 
 TEST_F(Bytecode0, Struct01) {
@@ -2379,6 +2157,447 @@ TEST_F(Bytecode0, Struct08) {
 
     const size_t stringssize = 0;
     EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Struct09) {
+
+    // Should be able to find SetCharacter as a component of
+    // VehicleBase as an extension of Vehicle Cars[5];
+    // should generate call of VehicleBase::SetCharacter()
+
+    char *inpl = "\
+        enum CharacterDirection                                     \n\
+        {                                                           \n\
+            eDirectionUp = 3                                        \n\
+        };                                                          \n\
+                                                                    \n\
+        builtin managed struct Character                            \n\
+        {                                                           \n\
+            readonly import attribute int ID;                       \n\
+        };                                                          \n\
+        import Character character[7];                              \n\
+        import Character cAICar1;                                   \n\
+                                                                    \n\
+        struct VehicleBase                                          \n\
+        {                                                           \n\
+            import void SetCharacter(Character *c,                  \n\
+                                int carSprite,                      \n\
+                                CharacterDirection carSpriteDir,    \n\
+                                int view = 0,                       \n\
+                                int loop = 0,                       \n\
+                                int frame = 0);                     \n\
+        };                                                          \n\
+                                                                    \n\
+        struct Vehicle extends VehicleBase                          \n\
+        {                                                           \n\
+            float bodyMass;                                         \n\
+        };                                                          \n\
+        import Vehicle Cars[6];                                     \n\
+                                                                    \n\
+        int main()                                                  \n\
+        {                                                           \n\
+            int drivers[] = new int[6];                             \n\
+            int i = 5;                                              \n\
+            Cars[i].SetCharacter(                                   \n\
+                character[cAICar1.ID + i],                          \n\
+                7 + drivers[i],                                     \n\
+                eDirectionUp,                                       \n\
+                3 + i, 0, 0);                                       \n\
+        }                                                           \n\
+    ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+    // WriteOutput("Struct09", scrip);
+    const size_t codesize = 193;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,            6,   72,    3,    4,    // 7
+       0,   51,    0,   47,            3,    1,    1,    4,    // 15
+       6,    3,    5,   29,            3,    6,    2,    4,    // 23
+      29,    2,   51,    8,            7,    3,   30,    2,    // 31
+      46,    3,    6,   32,            3,    4,   11,    2,    // 39
+       3,   29,    2,    6,            3,    0,   34,    3,    // 47
+       6,    3,    0,   34,            3,    6,    3,    3,    // 55
+      29,    3,   51,   12,            7,    3,   30,    4,    // 63
+      11,    4,    3,    3,            4,    3,   34,    3,    // 71
+       6,    3,    3,   34,            3,    6,    3,    7,    // 79
+      29,    3,   51,   16,           48,    2,   52,   29,    // 87
+       2,   51,   16,    7,            3,   30,    2,   32,    // 95
+       3,    4,   71,    3,           11,    2,    3,    7,    // 103
+       3,   30,    4,   11,            4,    3,    3,    4,    // 111
+       3,   34,    3,    6,            2,    1,   29,    2,    // 119
+       6,    2,    2,   29,            6,   45,    2,   39,    // 127
+       0,    6,    3,    0,           33,    3,   30,    6,    // 135
+      29,    3,   51,   16,            7,    3,   30,    4,    // 143
+      11,    4,    3,    3,            4,    3,   30,    2,    // 151
+      46,    3,    7,   32,            3,    0,   11,    2,    // 159
+       3,    3,    2,    3,           34,    3,   51,    4,    // 167
+       7,    2,   45,    2,           39,    6,    6,    3,    // 175
+       3,   33,    3,   35,            6,   30,    2,   51,    // 183
+       8,   49,    2,    1,            8,    6,    3,    0,    // 191
+       5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 5;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+      23,  117,  122,  131,        176,  -999
+    };
+    char fixuptypes[] = {
+      4,   4,   4,   4,      4,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 5;
+    std::string imports[] = {
+    "Character::get_ID^0",        "character",   "cAICar1",     "VehicleBase::SetCharacter^6",               // 3
+    "Cars",         "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Struct10) {
+
+    // When accessing a component of an import variable,
+    // the import variable must be read first so that the fixup can be
+    // applied. Only then may the offset be added to it.
+
+    char *inpl = "\
+        import struct Struct                                 \n\
+        {                                                    \n\
+            int fluff;                                       \n\
+            int k;                                           \n\
+        } ss;                                                \n\
+                                                             \n\
+        int main()                                           \n\
+        {                                                    \n\
+            return ss.k;                                     \n\
+        }                                                    \n\
+    ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Struct10", scrip);
+    const size_t codesize = 16;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    2,            0,    1,    2,    4,    // 7
+       7,    3,   31,    3,            6,    3,    0,    5,    // 15
+     -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 1;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+       4,  -999
+    };
+    char fixuptypes[] = {
+      4,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 1;
+    std::string imports[] = {
+    "ss",           "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Struct11) {
+
+    // Structs may contain variables that are structs themselves.
+    // Since Inner1 is managed, In1 will convert into an Inner1 *.
+
+    char *inpl = "\
+        managed struct Inner1                               \n\
+        {                                                   \n\
+            short Fluff;                                    \n\
+            int Payload;                                    \n\
+        };                                                  \n\
+        struct Inner2                                       \n\
+        {                                                   \n\
+            short Fluff;                                    \n\
+            int Payload;                                    \n\
+        };                                                  \n\
+        import int Foo;                                     \n\
+        import struct Struct                                \n\
+        {                                                   \n\
+            Inner1 In1;                                     \n\
+            Inner2 In2;                                     \n\
+        } SS;                                               \n\
+                                                            \n\
+        int main()                                          \n\
+        {                                                   \n\
+            SS.In1 = new Inner1;                            \n\
+            SS.In1.Payload = 77;                            \n\
+            SS.In2.Payload = 777;                           \n\
+            return SS.In1.Payload + SS.In2.Payload;         \n\
+        }                                                   \n\
+    ";
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Struct11", scrip);
+    const size_t codesize = 78;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   73,    3,            8,    6,    2,    1,    // 7
+      47,    3,    6,    3,           77,   29,    3,    6,    // 15
+       2,    1,   48,    2,           52,    1,    2,    2,    // 23
+      30,    3,    8,    3,            6,    3,  777,   29,    // 31
+       3,    6,    2,    1,            1,    2,    6,   30,    // 39
+       3,    8,    3,    6,            2,    1,   48,    2,    // 47
+      52,    1,    2,    2,            7,    3,   29,    3,    // 55
+       6,    2,    1,    1,            2,    6,    7,    3,    // 63
+      30,    4,   11,    4,            3,    3,    4,    3,    // 71
+      31,    3,    6,    3,            0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 5;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+       7,   17,   35,   45,         58,  -999
+    };
+    char fixuptypes[] = {
+      4,   4,   4,   4,      4,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 1;
+    std::string imports[] = {
+    "SS",           "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Struct12) {
+
+    // Can have managed components in non-managed struct.
+
+    char *inpl = "\
+        struct NonManaged           \n\
+        {                           \n\
+            long Dummy;             \n\
+            int  IntArray[];        \n\
+        } SS;                       \n\
+                                    \n\
+        int main()                  \n\
+        {                           \n\
+            SS.IntArray = new int[10];  \n\
+            SS.IntArray[3] = 7;     \n\
+            return SS.IntArray[3];  \n\
+        }                           \n\
+        ";
+
+    int compileResult = cc_compile(inpl, scrip);
+
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    WriteOutput("Struct12", scrip);
+
+    size_t const codesize = 49;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,           10,   72,    3,    4,    // 7
+       0,    6,    2,    4,           47,    3,    6,    3,    // 15
+       7,   29,    3,    6,            2,    4,   48,    2,    // 23
+      52,    1,    2,   12,           30,    3,    8,    3,    // 31
+       6,    2,    4,   48,            2,   52,    1,    2,    // 39
+      12,    7,    3,   31,            3,    6,    3,    0,    // 47
+       5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    size_t const numfixups = 3;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+      11,   21,   34,  -999
+    };
+    char fixuptypes[] = {
+      1,   1,   1,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    int const numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    size_t const numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    size_t const stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, StructArray01) {
+
+    char *inpl = "\
+    managed struct Struct                \n\
+    {                                    \n\
+        int Int[10];                     \n\
+    };                                   \n\
+                                         \n\
+    int main()                           \n\
+    {                                    \n\
+        Struct *S = new Struct;          \n\
+        S.Int[4] =  1;                   \n\
+    }                                    \n\
+    ";
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("StructArray01", scrip);
+    const size_t codesize = 39;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   73,    3,           40,   51,    0,   47,    // 7
+       3,    1,    1,    4,            6,    3,    1,   29,    // 15
+       3,   51,    8,   48,            2,   52,    1,    2,    // 23
+      16,   30,    3,    8,            3,   51,    4,   49,    // 31
+       2,    1,    4,    6,            3,    0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, StructArray02) {
+
+    // Static arrays can be multidimensional
+
+    char *inpl = "\
+    managed struct Struct                \n\
+    {                                    \n\
+        int Int1[5, 4];                  \n\
+        int Int2[2][3];                  \n\
+    };                                   \n\
+                                         \n\
+    int main()                           \n\
+    {                                    \n\
+        Struct S = new Struct;           \n\
+        S.Int1[4, 2] = 1;                \n\
+        S.Int2[1][2] = S.Int1[4, 2];     \n\
+    }                                    \n\
+    ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("StructArray02", scrip);
+    const size_t codesize = 63;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   73,    3,          104,   51,    0,   47,    // 7
+       3,    1,    1,    4,            6,    3,    1,   29,    // 15
+       3,   51,    8,   48,            2,   52,    1,    2,    // 23
+      72,   30,    3,    8,            3,   51,    4,   48,    // 31
+       2,   52,    1,    2,           72,    7,    3,   29,    // 39
+       3,   51,    8,   48,            2,   52,    1,    2,    // 47
+     100,   30,    3,    8,            3,   51,    4,   49,    // 55
+       2,    1,    4,    6,            3,    0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func00) {
+
+    char *inpl = "\
+        int Foo()      \n\
+    {                  \n\
+        return 15;     \n\
+    }";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func00", scrip);
+    const size_t codesize = 11;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,           15,   31,    3,    6,    // 7
+       3,    0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+
 }
 
 TEST_F(Bytecode0, Func01) {
@@ -3002,6 +3221,457 @@ TEST_F(Bytecode0, Func10) {
     EXPECT_EQ(stringssize, scrip.stringssize);
 }
 
+TEST_F(Bytecode0, Func11) {
+
+    char *inpl = "\
+    int Func(int I, ...)                 \n\
+    {                                    \n\
+        return I + I / I;                \n\
+    }                                    \n\
+                                         \n\
+    int main()                           \n\
+    {                                    \n\
+        return 0;                        \n\
+    }                                    \n\
+    ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func11", scrip);
+    const size_t codesize = 51;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   51,    8,            7,    3,   29,    3,    // 7
+      51,   12,    7,    3,           29,    3,   51,   16,    // 15
+       7,    3,   30,    4,           10,    4,    3,    3,    // 23
+       4,    3,   30,    4,           11,    4,    3,    3,    // 31
+       4,    3,   31,    3,            6,    3,    0,    5,    // 39
+      38,   40,    6,    3,            0,   31,    3,    6,    // 47
+       3,    0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func12) {
+
+    // Function with float default, or default "0", for float parameter
+
+    char *inpl = "\
+    float Func1(float F = 7.2)          \n\
+    {                                   \n\
+        return F;                       \n\
+    }                                   \n\
+                                        \n\
+    float Func2(float F = 0)            \n\
+    {                                   \n\
+        return F;                       \n\
+    }                                   \n\
+                                        \n\
+    float Call()                        \n\
+    {                                   \n\
+        return Func1() + Func2();       \n\
+    }                                   \n\
+    ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func12", scrip);
+    const size_t codesize = 68;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   51,    8,            7,    3,   31,    3,    // 7
+       6,    3,    0,    5,           38,   12,   51,    8,    // 15
+       7,    3,   31,    3,            6,    3,    0,    5,    // 23
+      38,   24,    6,    3,         1088841318,   29,    3,    6,    // 31
+       3,    0,   23,    3,            2,    1,    4,   29,    // 39
+       3,    6,    3,    0,           29,    3,    6,    3,    // 47
+      12,   23,    3,    2,            1,    4,   30,    4,    // 55
+      57,    4,    3,    3,            4,    3,   31,    3,    // 63
+       6,    3,    0,    5,          -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 2;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+      33,   48,  -999
+    };
+    char fixuptypes[] = {
+      2,   2,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func13) {
+
+    // Function with default null or 0 for managed parameter
+
+    char *inpl = "\
+    managed struct S                    \n\
+    {                                   \n\
+        float f;                        \n\
+    };                                  \n\
+                                        \n\
+    S *Func1(S s = null)                \n\
+    {                                   \n\
+        return s;                       \n\
+    }                                   \n\
+                                        \n\
+    S *Func2(S s = 0)                   \n\
+    {                                   \n\
+        return s;                       \n\
+    }                                   \n\
+                                        \n\
+    int Call()                           \n\
+    {                                   \n\
+        return Func1() == Func2();      \n\
+    }                                   \n\
+    ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func13", scrip);
+    const size_t codesize = 118;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   51,    8,            7,    3,   50,    3,    // 7
+      51,    8,   48,    3,           29,    3,   51,    4,    // 15
+      50,    3,   51,   12,           49,   51,    4,   48,    // 23
+       3,   69,   30,    4,           31,    6,   51,    8,    // 31
+      49,    6,    3,    0,            5,   38,   37,   51,    // 39
+       8,    7,    3,   50,            3,   51,    8,   48,    // 47
+       3,   29,    3,   51,            4,   50,    3,   51,    // 55
+      12,   49,   51,    4,           48,    3,   69,   30,    // 63
+       4,   31,    6,   51,            8,   49,    6,    3,    // 71
+       0,    5,   38,   74,            6,    3,    0,   29,    // 79
+       3,    6,    3,    0,           23,    3,    2,    1,    // 87
+       4,   29,    3,    6,            3,    0,   29,    3,    // 95
+       6,    3,   37,   23,            3,    2,    1,    4,    // 103
+      30,    4,   15,    4,            3,    3,    4,    3,    // 111
+      31,    3,    6,    3,            0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 2;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+      83,   98,  -999
+    };
+    char fixuptypes[] = {
+      2,   2,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func14) {
+
+    // Strange misalignment due to bad function protocol
+
+    char *inpl = "\
+        struct Struct               \n\
+        {                           \n\
+            int A[];                \n\
+            int B[];                \n\
+                                    \n\
+            import void Test(int Arg);  \n\
+        };                          \n\
+                                    \n\
+        void Struct::Test(int Arg)  \n\
+        {                           \n\
+            this.A = new int[1];    \n\
+            this.B = new int[1];    \n\
+            this.B[0] = 123;        \n\
+            Display(this.A[0], this.B[0]); \n\
+        }                           \n\
+                                    \n\
+        void Display(int X, int Y)  \n\
+        {                           \n\
+        }                           \n\
+                                    \n\
+        int main()                  \n\
+        {                           \n\
+            Struct TS;              \n\
+            TS.Test(7);             \n\
+        }                           \n\
+        ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func14", scrip);
+    const size_t codesize = 139;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,            1,   72,    3,    4,    // 7
+       0,    3,    6,    2,           52,   47,    3,    6,    // 15
+       3,    1,   72,    3,            4,    0,    3,    6,    // 23
+       2,   52,    1,    2,            4,   47,    3,    6,    // 31
+       3,  123,   29,    3,            3,    6,    2,   52,    // 39
+       1,    2,    4,   48,            2,   52,   30,    3,    // 47
+       8,    3,   29,    6,            3,    6,    2,   52,    // 55
+       1,    2,    4,   48,            2,   52,    7,    3,    // 63
+      29,    3,    3,    6,            2,   52,   48,    2,    // 71
+      52,    7,    3,   29,            3,    6,    3,   88,    // 79
+      23,    3,    2,    1,            8,   30,    6,    5,    // 87
+      38,   88,    5,   38,           91,   51,    0,   63,    // 95
+       8,    1,    1,    8,           51,    8,   29,    2,    // 103
+       6,    3,    7,   29,            3,   51,    8,    7,    // 111
+       2,   45,    2,    6,            3,    0,   23,    3,    // 119
+       2,    1,    4,   30,            2,   51,    8,   49,    // 127
+       1,    2,    4,   49,            2,    1,    8,    6,    // 135
+       3,    0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 2;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+      79,  117,  -999
+    };
+    char fixuptypes[] = {
+      2,   2,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func15) {
+
+    char *inpl = "\
+        int a = 15;    \n\
+        int Foo( )     \n\
+        {              \n\
+            return a;  \n\
+        }";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func15", scrip);
+    const size_t codesize = 13;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    2,            0,    7,    3,   31,    // 7
+       3,    6,    3,    0,            5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 1;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+       4,  -999
+    };
+    char fixuptypes[] = {
+      1,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func16) {
+
+    char *inpl = "\
+        int Foo(int a) \n\
+    {                  \n\
+        return a;      \n\
+    }";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func16", scrip);
+    const size_t codesize = 12;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,   51,    8,            7,    3,   31,    3,    // 7
+       6,    3,    0,    5,          -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, Func17) {
+
+    char *inpl = "\
+        int Foo()       \n\
+        {               \n\
+            int a = 15; \n\
+            return a;   \n\
+        }";
+
+    int compileResult = cc_compile(inpl, scrip);
+    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("Func17", scrip);
+    const size_t codesize = 23;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,           15,   29,    3,   51,    // 7
+       4,    7,    3,    2,            1,    4,   31,    6,    // 15
+       2,    1,    4,    6,            3,    0,    5,  -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 0;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
+TEST_F(Bytecode0, FuncStart1) {
+
+    // NON-managed dynpointers must be read/rewritten at function start, too.
+
+    char inpl[] = "\
+        int Random(int X)                       \n\
+        {                                       \n\
+            Shuffle(new int[15], 10);           \n\
+        }                                       \n\
+        void Shuffle(int Ints[], int Length)    \n\
+        {                                       \n\
+        }                                       \n\
+        ";
+
+    int compileResult = cc_compile(inpl, scrip);
+    EXPECT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
+
+    // WriteOutput("FuncStart1", scrip);
+    const size_t codesize = 40;
+    EXPECT_EQ(codesize, scrip.codesize);
+
+    int32_t code[] = {
+      38,    0,    6,    3,           10,   29,    3,    6,    // 7
+       3,   15,   72,    3,            4,    0,   29,    3,    // 15
+       6,    3,   28,   23,            3,    2,    1,    8,    // 23
+       6,    3,    0,    5,           38,   28,   51,    8,    // 31
+       7,    3,   50,    3,           51,    8,   49,    5,    // 39
+     -999
+    };
+    CompareCode(&scrip, codesize, code);
+
+    const size_t numfixups = 1;
+    EXPECT_EQ(numfixups, scrip.numfixups);
+
+    int32_t fixups[] = {
+      18,  -999
+    };
+    char fixuptypes[] = {
+      2,  '\0'
+    };
+    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
+
+    const int numimports = 0;
+    std::string imports[] = {
+     "[[SENTINEL]]"
+    };
+    CompareImports(&scrip, numimports, imports);
+
+    const size_t numexports = 0;
+    EXPECT_EQ(numexports, scrip.numexports);
+
+    const size_t stringssize = 0;
+    EXPECT_EQ(stringssize, scrip.stringssize);
+}
+
 TEST_F(Bytecode0, Export) {
     
     char *inpl = "\
@@ -3190,435 +3860,6 @@ TEST_F(Bytecode0, ArrayOfPointers2) {
 
     const size_t numfixups = 0;
     EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, ArrayInStruct1) {
-
-    char *inpl = "\
-    managed struct Struct                \n\
-    {                                    \n\
-        int Int[10];                     \n\
-    };                                   \n\
-                                         \n\
-    int main()                           \n\
-    {                                    \n\
-        Struct *S = new Struct;          \n\
-        S.Int[4] =  1;                   \n\
-    }                                    \n\
-    ";
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("ArrayInStruct1", scrip);
-    const size_t codesize = 39;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   73,    3,           40,   51,    0,   47,    // 7
-       3,    1,    1,    4,            6,    3,    1,   29,    // 15
-       3,   51,    8,   48,            2,   52,    1,    2,    // 23
-      16,   30,    3,    8,            3,   51,    4,   49,    // 31
-       2,    1,    4,    6,            3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, ArrayInStruct2) {
-    
-    // Static arrays can be multidimensional
-
-    char *inpl = "\
-    managed struct Struct                \n\
-    {                                    \n\
-        int Int1[5, 4];                  \n\
-        int Int2[2][3];                  \n\
-    };                                   \n\
-                                         \n\
-    int main()                           \n\
-    {                                    \n\
-        Struct S = new Struct;           \n\
-        S.Int1[4, 2] = 1;                \n\
-        S.Int2[1][2] = S.Int1[4, 2];     \n\
-    }                                    \n\
-    ";
-    
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("ArrayInStruct2", scrip);
-    const size_t codesize = 63;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   73,    3,          104,   51,    0,   47,    // 7
-       3,    1,    1,    4,            6,    3,    1,   29,    // 15
-       3,   51,    8,   48,            2,   52,    1,    2,    // 23
-      72,   30,    3,    8,            3,   51,    4,   48,    // 31
-       2,   52,    1,    2,           72,    7,    3,   29,    // 39
-       3,   51,    8,   48,            2,   52,    1,    2,    // 47
-     100,   30,    3,    8,            3,   51,    4,   49,    // 55
-       2,    1,    4,    6,            3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Func11) { 
-
-    char *inpl = "\
-    int Func(int I, ...)                 \n\
-    {                                    \n\
-        return I + I / I;                \n\
-    }                                    \n\
-                                         \n\
-    int main()                           \n\
-    {                                    \n\
-        return 0;                        \n\
-    }                                    \n\
-    ";
-  
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Func11", scrip);
-    const size_t codesize = 51;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   51,    8,            7,    3,   29,    3,    // 7
-      51,   12,    7,    3,           29,    3,   51,   16,    // 15
-       7,    3,   30,    4,           10,    4,    3,    3,    // 23
-       4,    3,   30,    4,           11,    4,    3,    3,    // 31
-       4,    3,   31,    3,            6,    3,    0,    5,    // 39
-      38,   40,    6,    3,            0,   31,    3,    6,    // 47
-       3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Func12) {
-
-    // Function with float default, or default "0", for float parameter
-
-    char *inpl = "\
-    float Func1(float F = 7.2)          \n\
-    {                                   \n\
-        return F;                       \n\
-    }                                   \n\
-                                        \n\
-    float Func2(float F = 0)            \n\
-    {                                   \n\
-        return F;                       \n\
-    }                                   \n\
-                                        \n\
-    float Call()                        \n\
-    {                                   \n\
-        return Func1() + Func2();       \n\
-    }                                   \n\
-    ";
-
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Func12", scrip);
-    const size_t codesize = 68;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   51,    8,            7,    3,   31,    3,    // 7
-       6,    3,    0,    5,           38,   12,   51,    8,    // 15
-       7,    3,   31,    3,            6,    3,    0,    5,    // 23
-      38,   24,    6,    3,         1088841318,   29,    3,    6,    // 31
-       3,    0,   23,    3,            2,    1,    4,   29,    // 39
-       3,    6,    3,    0,           29,    3,    6,    3,    // 47
-      12,   23,    3,    2,            1,    4,   30,    4,    // 55
-      57,    4,    3,    3,            4,    3,   31,    3,    // 63
-       6,    3,    0,    5,          -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      33,   48,  -999
-    };
-    char fixuptypes[] = {
-      2,   2,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Func13) {
-
-    // Function with default null or 0 for managed parameter
-
-    char *inpl = "\
-    managed struct S                    \n\
-    {                                   \n\
-        float f;                        \n\
-    };                                  \n\
-                                        \n\
-    S *Func1(S s = null)                \n\
-    {                                   \n\
-        return s;                       \n\
-    }                                   \n\
-                                        \n\
-    S *Func2(S s = 0)                   \n\
-    {                                   \n\
-        return s;                       \n\
-    }                                   \n\
-                                        \n\
-    int Call()                           \n\
-    {                                   \n\
-        return Func1() == Func2();      \n\
-    }                                   \n\
-    ";
-   
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Func13", scrip);
-    const size_t codesize = 118;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   51,    8,            7,    3,   50,    3,    // 7
-      51,    8,   48,    3,           29,    3,   51,    4,    // 15
-      50,    3,   51,   12,           49,   51,    4,   48,    // 23
-       3,   69,   30,    4,           31,    6,   51,    8,    // 31
-      49,    6,    3,    0,            5,   38,   37,   51,    // 39
-       8,    7,    3,   50,            3,   51,    8,   48,    // 47
-       3,   29,    3,   51,            4,   50,    3,   51,    // 55
-      12,   49,   51,    4,           48,    3,   69,   30,    // 63
-       4,   31,    6,   51,            8,   49,    6,    3,    // 71
-       0,    5,   38,   74,            6,    3,    0,   29,    // 79
-       3,    6,    3,    0,           23,    3,    2,    1,    // 87
-       4,   29,    3,    6,            3,    0,   29,    3,    // 95
-       6,    3,   37,   23,            3,    2,    1,    4,    // 103
-      30,    4,   15,    4,            3,    3,    4,    3,    // 111
-      31,    3,    6,    3,            0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      83,   98,  -999
-    };
-    char fixuptypes[] = {
-      2,   2,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Func14) {   
-
-    // Strange misalignment due to bad function protocol
-
-    char *inpl = "\
-        struct Struct               \n\
-        {                           \n\
-            int A[];                \n\
-            int B[];                \n\
-                                    \n\
-            import void Test(int Arg);  \n\
-        };                          \n\
-                                    \n\
-        void Struct::Test(int Arg)  \n\
-        {                           \n\
-            this.A = new int[1];    \n\
-            this.B = new int[1];    \n\
-            this.B[0] = 123;        \n\
-            Display(this.A[0], this.B[0]); \n\
-        }                           \n\
-                                    \n\
-        void Display(int X, int Y)  \n\
-        {                           \n\
-        }                           \n\
-                                    \n\
-        int main()                  \n\
-        {                           \n\
-            Struct TS;              \n\
-            TS.Test(7);             \n\
-        }                           \n\
-        ";
-    
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Func14", scrip);
-    const size_t codesize = 139;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,            1,   72,    3,    4,    // 7
-       0,    3,    6,    2,           52,   47,    3,    6,    // 15
-       3,    1,   72,    3,            4,    0,    3,    6,    // 23
-       2,   52,    1,    2,            4,   47,    3,    6,    // 31
-       3,  123,   29,    3,            3,    6,    2,   52,    // 39
-       1,    2,    4,   48,            2,   52,   30,    3,    // 47
-       8,    3,   29,    6,            3,    6,    2,   52,    // 55
-       1,    2,    4,   48,            2,   52,    7,    3,    // 63
-      29,    3,    3,    6,            2,   52,   48,    2,    // 71
-      52,    7,    3,   29,            3,    6,    3,   88,    // 79
-      23,    3,    2,    1,            8,   30,    6,    5,    // 87
-      38,   88,    5,   38,           91,   51,    0,   63,    // 95
-       8,    1,    1,    8,           51,    8,   29,    2,    // 103
-       6,    3,    7,   29,            3,   51,    8,    7,    // 111
-       2,   45,    2,    6,            3,    0,   23,    3,    // 119
-       2,    1,    4,   30,            2,   51,    8,   49,    // 127
-       1,    2,    4,   49,            2,    1,    8,    6,    // 135
-       3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      79,  117,  -999
-    };
-    char fixuptypes[] = {
-      2,   2,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, FuncStart1) {
-
-    // NON-managed dynpointers must be read/rewritten at function start, too.
-
-    char inpl[] = "\
-        int Random(int X)                       \n\
-        {                                       \n\
-            Shuffle(new int[15], 10);           \n\
-        }                                       \n\
-        void Shuffle(int Ints[], int Length)    \n\
-        {                                       \n\
-        }                                       \n\
-        ";
-    
-    int compileResult = cc_compile(inpl, scrip);
-    EXPECT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("FuncStart1", scrip);
-    const size_t codesize = 40;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,           10,   29,    3,    6,    // 7
-       3,   15,   72,    3,            4,    0,   29,    3,    // 15
-       6,    3,   28,   23,            3,    2,    1,    8,    // 23
-       6,    3,    0,    5,           38,   28,   51,    8,    // 31
-       7,    3,   50,    3,           51,    8,   49,    5,    // 39
-     -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 1;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      18,  -999
-    };
-    char fixuptypes[] = {
-      2,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
 
     const int numimports = 0;
     std::string imports[] = {
@@ -3989,370 +4230,4 @@ TEST_F(Bytecode0, Import) {
 
     const size_t stringssize = 0;
     EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Switch02) {
-    
-    // Last switch clause no "break"
-    char *inpl = "\
-        void main()                     \n\
-        {                               \n\
-            int i = 5;                  \n\
-            switch(i)                   \n\
-            {                           \n\
-            default: break;             \n\
-            case 5: i = 0;              \n\
-            }                           \n\
-            return;                     \n\
-        }                               \n\
-        ";
-   
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Switch02", scrip);
-    const size_t codesize = 50;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,            5,   29,    3,   51,    // 7
-       4,    7,    3,    3,            3,    4,   31,   11,    // 15
-      31,   23,    6,    3,            0,   51,    4,    8,    // 23
-       3,   31,   14,   29,            4,    6,    3,    5,    // 31
-      30,    4,   15,    3,            4,   70,  -21,   31,    // 39
-     -25,    2,    1,    4,           31,    3,    2,    1,    // 47
-       4,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 0;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-
-    int idx2 = -1;
-    for (size_t idx = 0; static_cast<int>(idx) < scrip.numimports; idx++)
-    {
-        if (!strcmp(scrip.imports[idx], ""))
-            continue;
-        idx2++;
-        ASSERT_LT(idx2, numimports);
-        std::string prefix = "imports[";
-        prefix += std::to_string(idx2) + "] == ";
-        std::string is_val = prefix + scrip.imports[idx];
-        std::string test_val = prefix + imports[idx2];
-        ASSERT_EQ(is_val, test_val);
-    }
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Attributes01) {    
-
-    char *inpl = "\
-        enum bool { false = 0, true = 1, };              \n\
-        builtin managed struct ViewFrame {              \n\
-            readonly import attribute bool Flipped;     \n\
-            import attribute int Graphic;               \n\
-            readonly import attribute float AsFloat;    \n\
-        };                                              \n\
-                                                        \n\
-        int main()                                      \n\
-        {                                               \n\
-            ViewFrame *VF;                              \n\
-            if (VF.Flipped)                             \n\
-            {                                           \n\
-                VF.Graphic = 17;                        \n\
-                float f = VF.AsFloat + VF.AsFloat;      \n\
-                return VF.Graphic;                      \n\
-            }                                           \n\
-            return VF.Flipped;                          \n\
-        }                                               \n\
-        ";
-
-    
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Attributes01", scrip);
-    const size_t codesize = 166;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   51,    0,           49,    1,    1,    4,    // 7
-      51,    4,   48,    2,           52,   29,    6,   45,    // 15
-       2,   39,    0,    6,            3,    0,   33,    3,    // 23
-      30,    6,   28,  102,            6,    3,   17,   51,    // 31
-       4,   48,    2,   52,           29,    6,   34,    3,    // 39
-      45,    2,   39,    1,            6,    3,    2,   33,    // 47
-       3,   35,    1,   30,            6,   51,    4,   48,    // 55
-       2,   52,   29,    6,           45,    2,   39,    0,    // 63
-       6,    3,    3,   33,            3,   30,    6,   29,    // 71
-       3,   51,    8,   48,            2,   52,   29,    6,    // 79
-      45,    2,   39,    0,            6,    3,    3,   33,    // 87
-       3,   30,    6,   30,            4,   57,    4,    3,    // 95
-       3,    4,    3,   29,            3,   51,    8,   48,    // 103
-       2,   52,   29,    6,           45,    2,   39,    0,    // 111
-       6,    3,    1,   33,            3,   30,    6,   51,    // 119
-       8,   49,    2,    1,            8,   31,   38,    2,    // 127
-       1,    4,   51,    4,           48,    2,   52,   29,    // 135
-       6,   45,    2,   39,            0,    6,    3,    0,    // 143
-      33,    3,   30,    6,           51,    4,   49,    2,    // 151
-       1,    4,   31,    9,           51,    4,   49,    2,    // 159
-       1,    4,    6,    3,            0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 6;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      21,   46,   66,   86,        114,  143,  -999
-    };
-    char fixuptypes[] = {
-      4,   4,   4,   4,      4,   4,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 4;
-    std::string imports[] = {
-    "ViewFrame::get_Flipped^0",   "ViewFrame::get_Graphic^0",   "ViewFrame::set_Graphic^1",   // 2
-    "ViewFrame::get_AsFloat^0",    "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Attributes02) {
-
-    // The getter and setter functions are defined locally, so
-    // they ought to be exported instead of imported.
-    // Assigning to the attribute should generate the same call
-    // as calling the setter; reading the same as calling the getter.
-    // Armor:: functions should be allowed to access _Damage.
-
-    char *inpl = "\
-        managed struct Armor {                          \n\
-            attribute int Damage;                       \n\
-            writeprotected short _Aura;                 \n\
-            protected int _Damage;                      \n\
-        };                                              \n\
-                                                        \n\
-        int main()                                      \n\
-        {                                               \n\
-            Armor *armor = new Armor;                   \n\
-            armor.Damage = 17;                          \n\
-            return (armor.Damage > 10);                 \n\
-        }                                               \n\
-                                                        \n\
-        void Armor::set_Damage(int damage)              \n\
-        {                                               \n\
-            if (damage >= 0)                            \n\
-                _Damage = damage;                       \n\
-        }                                               \n\
-                                                        \n\
-        int Armor::get_Damage()                         \n\
-        {                                               \n\
-            return _Damage;                             \n\
-        }                                               \n\
-        ";
-
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Attributes02", scrip);
-    const size_t codesize = 139;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   73,    3,            8,   51,    0,   47,    // 7
-       3,    1,    1,    4,            6,    3,   17,   51,    // 15
-       4,   48,    2,   52,           29,    6,   29,    3,    // 23
-      45,    2,    6,    3,           83,   23,    3,    2,    // 31
-       1,    4,   30,    6,           51,    4,   48,    2,    // 39
-      52,   29,    6,   45,            2,    6,    3,  122,    // 47
-      23,    3,   30,    6,           29,    3,    6,    3,    // 55
-      10,   30,    4,   17,            4,    3,    3,    4,    // 63
-       3,   51,    4,   49,            2,    1,    4,   31,    // 71
-       9,   51,    4,   49,            2,    1,    4,    6,    // 79
-       3,    0,    5,   38,           83,   51,    8,    7,    // 87
-       3,   29,    3,    6,            3,    0,   30,    4,    // 95
-      19,    4,    3,    3,            4,    3,   28,   17,    // 103
-      51,    8,    7,    3,           29,    3,    3,    6,    // 111
-       2,   52,    1,    2,            2,   30,    3,    8,    // 119
-       3,    5,   38,  122,            3,    6,    2,   52,    // 127
-       1,    2,    2,    7,            3,   31,    3,    6,    // 135
-       3,    0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      28,   47,  -999
-    };
-    char fixuptypes[] = {
-      2,   2,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, Attributes03) {
-
-    // The getters and setters are NOT defined locally, so
-    // import decls should be generated for them.
-    // The getters and setters should be called as import funcs.
-
-    char *inpl = "\
-        managed struct Armor {                          \n\
-            attribute int Damage;                       \n\
-            writeprotected short _aura;                 \n\
-            protected int _damage;                      \n\
-        };                                              \n\
-                                                        \n\
-        int main()                                      \n\
-        {                                               \n\
-            Armor *armor = new Armor;                   \n\
-            armor.Damage = 17;                          \n\
-            return (armor.Damage > 10);                 \n\
-        }";
-   
-    int compileResult = cc_compile(inpl, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("Attributes03", scrip);
-    const size_t codesize = 86;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,   73,    3,            8,   51,    0,   47,    // 7
-       3,    1,    1,    4,            6,    3,   17,   51,    // 15
-       4,   48,    2,   52,           29,    6,   34,    3,    // 23
-      45,    2,   39,    1,            6,    3,    1,   33,    // 31
-       3,   35,    1,   30,            6,   51,    4,   48,    // 39
-       2,   52,   29,    6,           45,    2,   39,    0,    // 47
-       6,    3,    0,   33,            3,   30,    6,   29,    // 55
-       3,    6,    3,   10,           30,    4,   17,    4,    // 63
-       3,    3,    4,    3,           51,    4,   49,    2,    // 71
-       1,    4,   31,    9,           51,    4,   49,    2,    // 79
-       1,    4,    6,    3,            0,    5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    const size_t numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-      30,   50,  -999
-    };
-    char fixuptypes[] = {
-      4,   4,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    const int numimports = 2;
-    std::string imports[] = {
-    "Armor::get_Damage^0",        "Armor::set_Damage^1",         "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    const size_t numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    const size_t stringssize = 0;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-}
-
-TEST_F(Bytecode0, StringStandard01) {    
-
-    char inpl[] = "\
-        int main()                         \n\
-        {                                  \n\
-            String s = \"Hello, world!\";  \n\
-            if (s != \"Bye\")              \n\
-                return 1;                  \n\
-            return 0;                      \n\
-        }                                  \n\
-        ";
-    std::string input = "";
-    input += g_Input_Bool;
-    input += g_Input_String;
-    input += inpl;
-
-    
-    int compileResult = cc_compile(input, scrip);
-    ASSERT_STREQ("Ok", (compileResult >= 0) ? "Ok" : last_seen_cc_error());
-
-    // WriteOutput("StringStandard01", scrip);
-    size_t const codesize = 65;
-    EXPECT_EQ(codesize, scrip.codesize);
-
-    int32_t code[] = {
-      38,    0,    6,    3,            0,   64,    3,   51,    // 7
-       0,   47,    3,    1,            1,    4,   51,    4,    // 15
-      48,    3,   29,    3,            6,    3,   14,   30,    // 23
-       4,   66,    4,    3,            3,    4,    3,   28,    // 31
-      11,    6,    3,    1,           51,    4,   49,    2,    // 39
-       1,    4,   31,   20,            6,    3,    0,   51,    // 47
-       4,   49,    2,    1,            4,   31,    9,   51,    // 55
-       4,   49,    2,    1,            4,    6,    3,    0,    // 63
-       5,  -999
-    };
-    CompareCode(&scrip, codesize, code);
-
-    size_t const numfixups = 2;
-    EXPECT_EQ(numfixups, scrip.numfixups);
-
-    int32_t fixups[] = {
-       4,   22,  -999
-    };
-
-    char fixuptypes[] = {
-      3,   3,  '\0'
-    };
-    CompareFixups(&scrip, numfixups, fixups, fixuptypes);
-
-    int const numimports = 0;
-    std::string imports[] = {
-     "[[SENTINEL]]"
-    };
-    CompareImports(&scrip, numimports, imports);
-
-    size_t const numexports = 0;
-    EXPECT_EQ(numexports, scrip.numexports);
-
-    size_t const stringssize = 18;
-    EXPECT_EQ(stringssize, scrip.stringssize);
-
-    char strings[] = {
-    'H',  'e',  'l',  'l',          'o',  ',',  ' ',  'w',     // 7
-    'o',  'r',  'l',  'd',          '!',    0,  'B',  'y',     // 15
-    'e',    0,  '\0'
-    };
-    CompareStrings(&scrip, stringssize, strings);
 }
